Fix interpolation artifacts
Occasionally, interpolation artifacts (rounding/precision errors) could
occur when a value was not changing between frames. This fix eliminates
that by approximating whether the values are "close enough" to be
equal- and thus make no change.

I was seeing this issue pop up for a sprite that looked to be twitching
when idling/not moving. It was kind of a cool effect, but not
intended...