Position vertices depending on pixel setting
Yikes. There's a good bunch of non-sense going on in here, and I feel
like i've dug myself into a hole with how precise this pixel stuff
should be/need to be (it worked just fine a revision ago...).

I *think* i've got it mostly right by this commit, but at the cost of
some half-pixel shifting that I can't quite explain every case of.

I'm certain there's a cleaner way of achieving this pixel precision
stuff so everything works as expected, while still being precise
(enough), but for now this will do.