Prefer applying fewer pixel adjustments
This commit addresses pixel accuracy once again. This time i've decided
to not worry about rasterization inaccuracies and instead deal with
those on the client side (i.e. making sure to position stuff correctly
and adding +1 pixel if necessary for the desired visual effect).

This does away with a lot of the complicated half-pixel adjustments we
previously had in place, making it simpler and easier to grasp.

So what's happening now is that you can still toggle flooring of
pixels, but it won't guarantee accurate rendering anymore; which is why
I personally prefer leaving it off.

Drawing points and filled rects now use the spritebatch instead of the
usual GL_POINTS/STRIPS. I feel like we might as well put more load on
that mechanism and make it the primary way of drawing stuff; this
results in two less drawing strategies to maintain.